requirements.txt
langchain
chromadb
pypdf
openai==0.26.1
python-dotenv
tiktoken